move exceptions to module and clean up linting errors
