+++ watcher
@@ -9,9 +9,9 @@ import time
 
 import click
 
+from nearuplib.exceptions import NetworkError
 from nearuplib.nodelib import restart_nearup
 from nearuplib.util import (
-    NetworkError,
     latest_deployed_release_commit,
     latest_deployed_release_commit_has_changed,
 )
@@ -55,7 +55,7 @@ def run(network, home, force_restart):
                 restart_nearup(network, home_dir=home)
                 current_release_commit = latest_deployed_release_commit(network)
 
-        except util.NetworkError as ex:
+        except NetworkError as ex:
             logging.warning(f'caught networking error {ex} - will try again')
 
         except Exception as ex:
